Crate pyrev_ast

Structs

Assert

断言

Assign

赋值

Attribute

属性 例如: a.b

BaseValue

String的Expression封装

BinaryOperation

二元操作 包括 +, -, *, /, <<, %, ==, >, is, in等

Call

函数调用

Container

容器(包括list, tuple, set, dict等)

Expr

只是对外提供一个ExpressionEnum的封装 (单纯不想使用Vec<ExpressionEnum>而已 )

Function

函数

If

If expression

Import

导入

Jump

Jump

QueryResMut

Raise

抛出异常

Return

返回

Slice

切片

UnaryOperation

Enums

ContainerType

ExpressionEnum

为上面的表达式提供一个封装 用来实现不同Expression的嵌套

UnaryType

Traits

Expression

Expression trait is used to mark the struct as an expression

Query

Query example:

QueryMutable

Queryable

Queryable trait is used to convert the struct to dyn Any It is used to search for the type T in the Struct

Derive Macros

Expression

Query

Query trait is depend on Queryable trait It searches for the type T recursively in the struct It will return immediately when if finds one, even if the return structures contains the target type T